* blackbox no longer uses every available screen by default.  If you want this   behavior pass -multi on the command line.  bsetroot still needs some work.
* a little work on Font.cc's ellideText() in preparation for its use on Window
  titles.

+++ lib/Font.hh
@@ -82,7 +82,7 @@ namespace bt {
                 const Rect &rect, Alignment alignment,
                 const std::string &text);
 
-  std::string ellideText(const std::string& text, unsigned int count,
+  std::string ellideText(const std::string& text, size_t count,
                          const char* ellide);
 
   Alignment alignResource(const Resource &resource,
